console: Refactor application views
