+++ pkg/webui/console/views/application-collaborator-add/index.js
@@ -1,4 +1,4 @@
-// Copyright © 2022 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2023 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,9 +12,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import ApplicationCollaboratorAdd from './application-collaborator-add'
-import connect from './connect'
+import React from 'react'
+import { Container, Col, Row } from 'react-grid-system'
+import { useParams } from 'react-router-dom'
 
-const ConnectedCollaboratorAdd = connect(ApplicationCollaboratorAdd)
+import { APPLICATION } from '@console/constants/entities'
 
-export { ConnectedCollaboratorAdd as default, ApplicationCollaboratorAdd }
+import PageTitle from '@ttn-lw/components/page-title'
+import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
+import { useBreadcrumbs } from '@ttn-lw/components/breadcrumbs/context'
+
+import ConsoleCollaboratorsForm from '@console/containers/collaborators-form'
+
+import sharedMessages from '@ttn-lw/lib/shared-messages'
+
+const ApplicationCollaboratorAdd = () => {
+  const { appId } = useParams()
+
+  useBreadcrumbs(
+    'apps.single.collaborators.add',
+    <Breadcrumb path={`/applications/${appId}/collaborators/add`} content={sharedMessages.add} />,
+  )
+
+  return (
+    <Container>
+      <PageTitle title={sharedMessages.addCollaborator} />
+      <Row>
+        <Col lg={8} md={12}>
+          <ConsoleCollaboratorsForm entity={APPLICATION} entityId={appId} />
+        </Col>
+      </Row>
+    </Container>
+  )
+}
+
+export default ApplicationCollaboratorAdd
