+++ pkg/webui/console/views/application-collaborator-edit/index.js
@@ -1,4 +1,4 @@
-// Copyright © 2022 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2023 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,9 +12,60 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import ApplicationCollaboratorEdit from './application-collaborator-edit'
-import connect from './connect'
+import React from 'react'
+import { Container, Col, Row } from 'react-grid-system'
+import { useParams } from 'react-router-dom'
 
-const ConnectedCollaboratorEdit = connect(ApplicationCollaboratorEdit)
+import { APPLICATION } from '@console/constants/entities'
 
-export { ConnectedCollaboratorEdit as default, ApplicationCollaboratorEdit }
+import PageTitle from '@ttn-lw/components/page-title'
+
+import RequireRequest from '@ttn-lw/lib/components/require-request'
+import GenericNotFound from '@ttn-lw/lib/components/full-view-error/not-found'
+
+import ConsoleCollaboratorsForm from '@console/containers/collaborators-form'
+
+import sharedMessages from '@ttn-lw/lib/shared-messages'
+import { getCollaborator, getCollaboratorsList } from '@ttn-lw/lib/store/actions/collaborators'
+
+const ApplicationCollaboratorEditInner = () => {
+  const { appId, collaboratorId } = useParams()
+
+  return (
+    <Container>
+      <PageTitle title={sharedMessages.collaboratorEdit} values={{ collaboratorId }} />
+      <Row>
+        <Col lg={8} md={12}>
+          <ConsoleCollaboratorsForm
+            entity={APPLICATION}
+            entityId={appId}
+            collaboratorId={collaboratorId}
+            collaboratorType="user"
+            update
+          />
+        </Col>
+      </Row>
+    </Container>
+  )
+}
+
+const ApplicationCollaboratorEdit = () => {
+  const { appId, collaboratorId, collaboratorType } = useParams()
+
+  if (collaboratorType !== 'user' && collaboratorType !== 'organization') {
+    return <GenericNotFound />
+  }
+
+  return (
+    <RequireRequest
+      requestAction={[
+        getCollaborator('application', appId, collaboratorId, collaboratorType === 'user'),
+        getCollaboratorsList('application', appId),
+      ]}
+    >
+      <ApplicationCollaboratorEditInner />
+    </RequireRequest>
+  )
+}
+
+export default ApplicationCollaboratorEdit
