+++ pkg/webui/console/views/application-integrations-webhook-edit/index.js
@@ -12,9 +12,113 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import ApplicationWebhookEdit from './application-integrations-webhook-edit'
-import connect from './connect'
+import React from 'react'
+import { Container, Col, Row } from 'react-grid-system'
+import { defineMessages } from 'react-intl'
+import { useSelector } from 'react-redux'
+import { useParams } from 'react-router-dom'
 
-const ConnectedApplicationWebhookEdit = connect(ApplicationWebhookEdit)
+import PageTitle from '@ttn-lw/components/page-title'
+import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
+import { useBreadcrumbs } from '@ttn-lw/components/breadcrumbs/context'
 
-export { ConnectedApplicationWebhookEdit as default, ApplicationWebhookEdit }
+import RequireRequest from '@ttn-lw/lib/components/require-request'
+
+import WebhookEdit from '@console/containers/webhook-edit'
+
+import Require from '@console/lib/components/require'
+
+import sharedMessages from '@ttn-lw/lib/shared-messages'
+import { getWebhookTemplateId } from '@ttn-lw/lib/selectors/id'
+
+import { getWebhook } from '@console/store/actions/webhooks'
+
+import { selectSelectedWebhook } from '@console/store/selectors/webhooks'
+import { selectWebhookTemplateById } from '@console/store/selectors/webhook-templates'
+import {
+  selectWebhooksHealthStatusEnabled,
+  selectWebhookRetryInterval,
+  selectWebhookHasUnhealthyConfig,
+} from '@console/store/selectors/application-server'
+
+const m = defineMessages({
+  editWebhook: 'Edit webhook',
+  updateSuccess: 'Webhook updated',
+  deleteSuccess: 'Webhook deleted',
+  reactivateSuccess: 'Webhook activated',
+})
+
+const ApplicationWebhookEditInner = () => {
+  const { appId, webhookId } = useParams()
+  const healthStatusEnabled = useSelector(selectWebhooksHealthStatusEnabled)
+  const webhookRetryInterval = useSelector(selectWebhookRetryInterval)
+  const hasUnhealthyWebhookConfig = useSelector(selectWebhookHasUnhealthyConfig)
+  const webhook = useSelector(selectSelectedWebhook)
+  const webhookTemplateId = getWebhookTemplateId(webhook)
+  const webhookTemplate = useSelector(state => selectWebhookTemplateById(state, webhookTemplateId))
+  useBreadcrumbs(
+    'apps.single.integrations.webhooks.edit',
+    <Breadcrumb
+      path={`/applications/${appId}/integrations/${webhookId}`}
+      content={sharedMessages.edit}
+    />,
+  )
+
+  return (
+    <Container>
+      <PageTitle title={m.editWebhook} className="mb-0" hideHeading={Boolean(webhookTemplate)} />
+      <Row>
+        <Col lg={8} md={12}>
+          <WebhookEdit
+            update
+            appId={appId}
+            selectedWebhook={webhook}
+            webhookId={webhookId}
+            webhookTemplate={webhookTemplate}
+            healthStatusEnabled={healthStatusEnabled}
+            webhookRetryInterval={webhookRetryInterval}
+            hasUnhealthyWebhookConfig={hasUnhealthyWebhookConfig}
+          />
+        </Col>
+      </Row>
+    </Container>
+  )
+}
+
+const webhookEntitySelector = [
+  'base_url',
+  'format',
+  'headers',
+  'downlink_api_key',
+  'uplink_message',
+  'uplink_normalized',
+  'join_accept',
+  'downlink_ack',
+  'downlink_nack',
+  'downlink_sent',
+  'downlink_failed',
+  'downlink_queued',
+  'downlink_queue_invalidated',
+  'location_solved',
+  'service_data',
+  'template_ids',
+  'health_status',
+  'field_mask',
+]
+
+const ApplicationWebhookEdit = () => {
+  const { appId, webhookId } = useParams()
+
+  const healthStatusEnabled = useSelector(selectWebhooksHealthStatusEnabled)
+  const hasUnhealthyWebhookConfig = useSelector(selectWebhookHasUnhealthyConfig)
+
+  return (
+    <Require featureCheck={healthStatusEnabled && !hasUnhealthyWebhookConfig}>
+      <RequireRequest requestAction={getWebhook(appId, webhookId, webhookEntitySelector)}>
+        <ApplicationWebhookEditInner />
+      </RequireRequest>
+    </Require>
+  )
+}
+
+export default ApplicationWebhookEdit
