+++ pkg/webui/console/views/applications/index.js
@@ -18,7 +18,8 @@ import { Routes, Route } from 'react-router-dom'
 import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
 import { useBreadcrumbs } from '@ttn-lw/components/breadcrumbs/context'
 
-import NotFoundRoute from '@ttn-lw/lib/components/not-found-route'
+import GenericNotFound from '@ttn-lw/lib/components/full-view-error/not-found'
+import ValidateRouteParam from '@ttn-lw/lib/components/validate-route-param'
 
 import withFeatureRequirement from '@console/lib/components/with-feature-requirement'
 
@@ -26,27 +27,24 @@ import Application from '@console/views/application'
 import ApplicationsList from '@console/views/applications-list'
 import ApplicationAdd from '@console/views/application-add'
 
-import PropTypes from '@ttn-lw/lib/prop-types'
 import sharedMessages from '@ttn-lw/lib/shared-messages'
 import { pathId as pathIdRegexp } from '@ttn-lw/lib/regexp'
 
 import { mayViewApplications } from '@console/lib/feature-checks'
 
-const Applications = props => {
-  const { path } = props.match
-
+const Applications = () => {
   useBreadcrumbs('apps', <Breadcrumb path="/applications" content={sharedMessages.applications} />)
 
   return (
     <Routes>
-      <Route exact path={`${path}`} component={ApplicationsList} />
-      <Route exact path={`${path}/add`} component={ApplicationAdd} />
-      <Route path={`${path}/:appId${pathIdRegexp}`} component={Application} sensitive />
-      <NotFoundRoute />
+      <Route index Component={ApplicationsList} />
+      <Route path="add" Component={ApplicationAdd} />
+      <Route
+        path=":appId/*"
+        element={<ValidateRouteParam check={{ appId: pathIdRegexp }} Component={Application} />}
+      />
+      <Route path="*" Component={GenericNotFound} />
     </Routes>
   )
-}
-Applications.propTypes = {
-  match: PropTypes.match.isRequired,
 }
 export default withFeatureRequirement(mayViewApplications, { redirect: '/' })(Applications)
